+++ packages/esm-appointments-app/src/hooks/useAppointments.ts
@@ -3,11 +3,36 @@ import dayjs from 'dayjs';
 import useSWR from 'swr';
 import { omrsDateFormat } from '../constants';
 import { type Appointment, DurationPeriod } from '../types';
+import { getAppointment, useAppointmentDate } from '../helpers';
+import isEmpty from 'lodash-es/isEmpty';
+import { useMemo } from 'react';
 
 interface AppointmentsReturnType {
   isLoading: boolean;
   appointments: Array<Appointment>;
   error: Error;
+}
+
+export function useAppointments(status?: string, forDate?: string) {
+  const { currentAppointmentDate } = useAppointmentDate();
+  const startDate = forDate ? forDate : currentAppointmentDate;
+  const apiUrl = `/ws/rest/v1/appointment/appointmentStatus?forDate=${startDate}&status=${status}`;
+  const allAppointmentsUrl = `/ws/rest/v1/appointment/all?forDate=${startDate}`;
+
+  const { data, error, isLoading, isValidating, mutate } = useSWR<{ data: Array<Appointment> }, Error>(
+    isEmpty(status) ? allAppointmentsUrl : apiUrl,
+    openmrsFetch,
+  );
+
+  const appointments = useMemo(() => data?.data?.map((appointment) => getAppointment(appointment)) ?? [], [data?.data]);
+
+  return {
+    appointments,
+    isLoading,
+    isError: error,
+    isValidating,
+    mutate,
+  };
 }
 
 export const useDailyAppointments = (startDateTime: string, durationPeriod: DurationPeriod) => {
