+++ packages/esm-appointments-app/src/root.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { SWRConfig } from 'swr';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import AppointmentsCalendarView from './appointments-calendar/appointments-calendar-view.component';
-import ClinicalAppointments from './appointments.component';
+import Appointments from './appointments.component';
 import CalendarPatientList from './appointments-calendar/patient-list/calendar-patient-list.component';
 
 const swrConfiguration = {
@@ -17,7 +17,7 @@ const RootComponent: React.FC = () => {
       <SWRConfig value={swrConfiguration}>
         <BrowserRouter basename={appointmentsBasename}>
           <Routes>
-            <Route path="/" element={<ClinicalAppointments />} />
+            <Route path="/" element={<Appointments />} />
             <Route path="/calendar" element={<AppointmentsCalendarView />} />
             <Route path="/list/:dateTime/:serviceName" element={<CalendarPatientList />} />
           </Routes>
